refactored devSpread method to reduce complexity and redundant loops
