+++ src/database/statistics/developerFocus.service.ts
@@ -5,13 +5,14 @@ import { RepositoryNameDto } from 'src/github-api/model/Repository';
 import {
   DevSpread,
   DevSpreadAvg,
-  DevSpreadDates,
   DevSpreadTotal,
   RepoSpread,
   RepoSpreadAvg,
   RepoSpreadTotal,
 } from 'src/github-api/model/PullRequest';
 import { RepositoryDocument } from '../schemas/repository.schema';
+import { rearangeTimeslots, spreadsGroupedByTimeslots } from './dateUtil';
+import { getRepoSpreadTotal, getAvgRepoSpread } from './spreadUtil';
 
 @Injectable()
 export class DeveloperFocus {
@@ -108,87 +109,89 @@ export class DeveloperFocus {
     loginFilter?: string[],
     userLimit?: number,
   ): Promise<DevSpread> {
-    // get all repo ids of the orga (which are already in the db)
-    const repoIds = await this.repoModel
+    const repoIds: { _id: string }[] = await this.repoModel
       .aggregate()
       .match({ owner: owner })
       .project({ _id: 1 })
       .exec();
 
-    // Store all repo commit objects (grouped by developer) in one object repoDevelopers.
-    // The repoId is the first key, then every repoId has developer props
-    // with all their commit timestamps array for that repo
-    const repoDevelopers: { [key: string]: { [key: string]: string[] } } = {};
-    const developerSet: Set<string> = new Set(); // store all unique developers
+    // devId -> TimeStamp -> RepoId
+    const devToTimestampToRepo = new Map<string, Map<string, string[]>>();
+
     for (const repoId of repoIds) {
-      const developers = await this.getRepoCommits(
+      const devToTimestamps = await this.getRepoCommits(
         repoId._id,
         loginFilter,
         userLimit,
       );
-      repoDevelopers[repoId._id] = developers;
-      for (const dev of Object.keys(developers)) {
-        developerSet.add(dev);
+
+      const repoIdString = repoId._id.toString();
+
+      for (const [dev, timestamps] of Object.entries(devToTimestamps)) {
+        if (!devToTimestampToRepo.has(dev)) {
+          this.addNewDevToMap(
+            dev,
+            timestamps,
+            repoIdString,
+            devToTimestampToRepo,
+          );
+        } else {
+          this.extendExistingDevEntry(
+            timestamps,
+            repoIdString,
+            devToTimestampToRepo.get(dev),
+          );
+        }
       }
     }
 
-    console.log(repoDevelopers);
-
-    // Now, construct an object spread with all developers as keys.
-    // Then, every developer has an object with timestamps as keys,
-    // and every timestamps has an array of RepoIds, so we achieve
-    // unique timestamps for each developer with grouped repos for that.
-    // Basically, this is the daySpread beeing calculated for each dev.
-    const spread: {
-      [key: string]: { [key: string]: [string] };
-    } = {};
-    // do this for all developers
-    for (const dev of developerSet) {
-      // allDevCommits stores all {timestamp:repo}[] pairs from all repos per each developer BEFORE grouping
-      const allDevCommits: { timestamp: string; repo: string }[] = [];
-      // do this for every repo
-      for (const repoId of Object.keys(repoDevelopers)) {
-        // do this for every dev in the repo, if it is the same as the current dev
-        for (const repoDev in repoDevelopers[repoId]) {
-          if (dev == repoDev) {
-            for (const timestamp of repoDevelopers[repoId][repoDev]) {
-              // store every timestamp:repoId pairs in allDevCommits
-              const commit = {
-                timestamp: timestamp,
-                repo: repoId,
-              };
-              allDevCommits.push(commit);
-            }
-          }
-        }
-      }
-      // directly group allDevCommits per key (timestamp)
-      // to get unique timestamps with an array of the corresponding repoID as value
-      spread[dev] = allDevCommits.reduce((acc, curr) => {
-        if (!acc[curr.timestamp]) {
-          acc[curr.timestamp] = [];
-        }
-        acc[curr.timestamp].push(curr.repo);
-        return acc;
-      }, {});
-    }
-    console.log(spread);
-
-    // Now, spreadsPerDevs (the final return value) is beeing calculated.
-    // Here, for every developer, all timestamps (or the begin-timestamp of an interval)
-    // are beeing stored with an array of corresponding repoIds precisely.
     const spreadsPerDevs: DevSpread = {};
 
-    // for every developer in the spread object,
-    // call helper function spreadsGroupedByTimeslots() to pass in
-    // the object with all timestamps as keys and repoId arrays as values,
-    // to compute all spreadsPerDevs props.
-    for (const dev of Object.keys(spread)) {
-      spreadsPerDevs[dev] = await this.spreadsGroupedByTimeslots(spread[dev]);
-    }
+    devToTimestampToRepo.forEach((timeToRepo, dev) => {
+      const timeToRepoObj = Object.fromEntries(timeToRepo);
+      // TODO: refactor spreadsGroupedByTimeslots
+      spreadsPerDevs[dev] = spreadsGroupedByTimeslots(timeToRepoObj);
+    });
 
     console.log(spreadsPerDevs);
     return spreadsPerDevs;
+  }
+
+  /**
+   *
+   * Adds the given to @param repoId to the timestamp map.
+   * If map doesn't contain the @param timestamps the entry is added.
+   */
+  private extendExistingDevEntry(
+    timestamps: string[],
+    repoId: string,
+    devToTimpestamps: Map<string, string[]>,
+  ) {
+    for (const timeStamp of timestamps) {
+      if (devToTimpestamps.has(timeStamp)) {
+        devToTimpestamps.get(timeStamp).push(repoId);
+      } else {
+        devToTimpestamps.set(timeStamp, [repoId]);
+      }
+    }
+  }
+
+  /**
+   * Adds a new map for every @param timpestamps to the
+   * given @param map, if the entry doesn't exist.
+   * The new map contains an array of repository ids.
+   */
+  private addNewDevToMap(
+    dev: string,
+    timestamps: string[],
+    repoId: string,
+    map: Map<string, Map<string, string[]>>,
+  ) {
+    const timeStampRepo = new Map<string, string[]>();
+    map.set(dev, timeStampRepo);
+    for (const timeStamp of timestamps) {
+      timeStampRepo.set(timeStamp, [repoId]);
+    }
   }
 
   /**
@@ -337,12 +340,14 @@ export class DeveloperFocus {
     const repoM = await this.repoModel
       .findOne({ repo: repoIdent.repo, owner: repoIdent.owner })
       .exec();
+    console.log(repoM);
     // get the commits for specified repo to get the developers of that repo
     const commits = await this.getRepoCommits(
       repoM._id,
       loginFilter,
       userLimit,
     );
+    console.log(commits);
     // store the repoId as a string
     const repoID = repoM._id.toString();
     // store the repo developers in an array
@@ -429,211 +434,24 @@ export class DeveloperFocus {
     // we want to have 2 weeks of them and add dev B to the week X and delete week Z.
     // This happens in rearangeTimeSlots() for weeks, sprints and months, as days
     // are already precisely.
-    dates.weekSpread = await this.rearangeTimeslots(dates.weekSpread, 7);
-    dates.sprintSpread = await this.rearangeTimeslots(dates.sprintSpread, 14);
-    dates.monthSpread = await this.rearangeTimeslots(dates.monthSpread, 30);
+    dates.weekSpread = rearangeTimeslots(dates.weekSpread, 7);
+    dates.sprintSpread = rearangeTimeslots(dates.sprintSpread, 14);
+    dates.monthSpread = rearangeTimeslots(dates.monthSpread, 30);
 
     console.log(dates);
 
     // this is the presicely repository spread with timestamp:spread pairs for each category
     // the spread is beeing calculated with the sum of the dev spreads which contributed in that timestamp
     // devided trough the amount of devs
-    const repoSpread: RepoSpreadTotal = {
-      daySpread: {},
-      weekSpread: {},
-      sprintSpread: {},
-      monthSpread: {},
-    };
-    for (const day in dates.daySpread) {
-      const daySpread =
-        Object.values(dates.daySpread[day]).reduce((a, b) => a + b) /
-        Object.values(dates.daySpread[day]).length;
-      repoSpread.daySpread[day] = daySpread;
-    }
-    for (const week in dates.weekSpread) {
-      const weekSpread =
-        Object.values(dates.weekSpread[week]).reduce((a, b) => a + b) /
-        Object.values(dates.weekSpread[week]).length;
-      repoSpread.weekSpread[week] = weekSpread;
-    }
-    for (const sprint in dates.sprintSpread) {
-      const sprintSpread =
-        Object.values(dates.sprintSpread[sprint]).reduce((a, b) => a + b) /
-        Object.values(dates.sprintSpread[sprint]).length;
-      repoSpread.sprintSpread[sprint] = sprintSpread;
-    }
-    for (const month in dates.monthSpread) {
-      const monthSpread =
-        Object.values(dates.monthSpread[month]).reduce((a, b) => a + b) /
-        Object.values(dates.monthSpread[month]).length;
-      repoSpread.monthSpread[month] = monthSpread;
-    }
+    const repoSpread: RepoSpreadTotal = getRepoSpreadTotal(dates);
 
     // TODO: this could also be returned as a good overview n the whole repo histroy, if desired.
     console.log(repoSpread);
 
     // finally, build the average daySpread, weekSpread, ..., of all single daySpreads, weekSpreads, ...
-    const avgRepoSpread: RepoSpreadAvg = {
-      daySpread:
-        Object.values(repoSpread.daySpread).reduce((a, b) => a + b) /
-        Object.values(repoSpread.daySpread).length,
-      weekSpread:
-        Object.values(repoSpread.weekSpread).reduce((a, b) => a + b) /
-        Object.values(repoSpread.weekSpread).length,
-      sprintSpread:
-        Object.values(repoSpread.sprintSpread).reduce((a, b) => a + b) /
-        Object.values(repoSpread.sprintSpread).length,
-      monthSpread:
-        Object.values(repoSpread.monthSpread).reduce((a, b) => a + b) /
-        Object.values(repoSpread.monthSpread).length,
-    };
+    const avgRepoSpread: RepoSpreadAvg = getAvgRepoSpread(repoSpread);
     console.log(avgRepoSpread);
 
     return avgRepoSpread;
-  }
-
-  /**
-   * Helper function to eliminate duplicate weeks, sprints or months
-   * because the calculated intervals per developer may differ
-   * and would have been taken as separate intervals into account,
-   * altough it was the same interval of commiting.
-   * @param timeSpreadPairs An object with timestamps as keys, and an object with developer:sprad pairs as a value
-   * @param days The days, the interval blongs to, i.e. 7 (weeks), 14 (sprint), 30(month)
-   * @returns
-   */
-  async rearangeTimeslots(
-    timeSpreadPairs: { [key: string]: { [key: string]: number } },
-    days: number,
-  ): Promise<{ [key: string]: { [key: string]: number } }> {
-    // copy the old timeSpreadPairs to modify on that
-    const newTimeSpreadPairs: { [key: string]: { [key: string]: number } } = {
-      ...timeSpreadPairs,
-    };
-    const timestamps: string[] = Object.keys(timeSpreadPairs).sort();
-    // go through every sorted timestamp and look to successor
-    for (let i = 1; i < timestamps.length; ) {
-      const currentDate: string = timestamps[i - 1];
-      const nextDate: string = timestamps[i];
-      // check, if the next interval date should actualy be in current interval
-      if ((await this.addDays(currentDate, days)) > new Date(timestamps[i])) {
-        // append the values together, update the current interval object
-        const devSpreadObj1 = timeSpreadPairs[currentDate];
-        const devSpreadObj2 = timeSpreadPairs[nextDate];
-        const mergedObj = { ...devSpreadObj1, ...devSpreadObj2 };
-        newTimeSpreadPairs[currentDate] = mergedObj;
-        // delete the unnessacary date and skip the next date
-        delete newTimeSpreadPairs[nextDate];
-        i += 2;
-      } else {
-        i += 1;
-      }
-    }
-    return newTimeSpreadPairs;
-  }
-
-  /**
-   * Helper function to add a number of days to an existing date.
-   * @param date The date, which should be increased.
-   * @param days The number of days, which should be added to the date.
-   * @returns The new increased date in Date() format
-   */
-  async addDays(date: string, days: number): Promise<Date> {
-    const result = new Date(date);
-    result.setDate(result.getDate() + days);
-    return result;
-  }
-
-  /**
-   * This function calculates all spread values for a given object with timestamps
-   * as keys and arrays of repoIds as values. Therefor, in daySpread, weekSpread,
-   * sprintSpread and monthSpread are stored all spreads precisely, i.e. their values
-   * considered as an object in which every timestamp (or beginning timestamp of that interval)
-   * is a key, with an array of the corresponding repoIds for that timestmamp as a value.
-   * The Sum properties are the sums of the lenghts of all repoId arrays in a category.
-   * The days, weeks, sprint, month properties are the amount of all keys (timestamps)
-   * in a category (for later avg computation).
-   * @param timeRepoPairs An Object of timestamps as keys, with repoId arrays as values (for a developer),
-   * i.e. it's already the daySpread of a developer.
-   * @returns An object with computed spreads for timeintervals, the sums of the single spreads
-   * and the number of timestamps for a category.
-   */
-  async spreadsGroupedByTimeslots(timeRepoPairs: {
-    [key: string]: [string];
-  }): Promise<DevSpreadDates> {
-    // sort the timestamps to ensure correct interval computation
-    const timestamps: string[] = Object.keys(timeRepoPairs).sort();
-    // the return object dates
-    const dates: DevSpreadDates = {
-      daySpread: {},
-      weekSpread: {},
-      sprintSpread: {},
-      monthSpread: {},
-      daySpreadSum: 0,
-      weekSpreadSum: 0,
-      sprintSpreadSum: 0,
-      monthSpreadSum: 0,
-      days: 0,
-      weeks: 0,
-      sprints: 0,
-      months: 0,
-    };
-    // init first dayspread entry
-    dates.daySpread[timestamps[0]] = timeRepoPairs[timestamps[0]];
-    // init all category counters with the first timestamp
-    let weekDate = timestamps[0];
-    let sprintDate = timestamps[0];
-    let monthDate = timestamps[0];
-    // counter for category sums
-    let daySpreadSum = timeRepoPairs[timestamps[0]].length; // already init it with the length of the first entrys repoID array
-    let weekSpreadSum = 0;
-    let sprintSpreadSum = 0;
-    let monthSpreadSum = 0;
-    // counstruct sets for week, sprint and month categorys to get unique repoIds to calculate the spreads
-    const weekSpread: Set<string> = new Set();
-    const sprintSpread: Set<string> = new Set();
-    const monthSpread: Set<string> = new Set();
-    // always compare the next date; if the interval fits, increase the category counters
-    for (let i = 1; i < timestamps.length; i++) {
-      // repo array for each timestamp
-      const repoArr = timeRepoPairs[timestamps[i]];
-      dates.daySpread[timestamps[i]] = repoArr; // always add entry to the daySpread category
-      daySpreadSum += timeRepoPairs[timestamps[i]].length; // always add the length (spread) to the daySpreadSum
-      // for other intervals, add every repo of current entry to the set
-      // Then, if one timeslot is reached, increase counters with length of the set,
-      // clear the set and set the category date to the new date
-      repoArr.forEach((repo) => {
-        weekSpread.add(repo);
-        sprintSpread.add(repo);
-        monthSpread.add(repo);
-      });
-      if ((await this.addDays(weekDate, 7)) <= new Date(timestamps[i])) {
-        dates.weekSpread[weekDate] = Array.from(weekSpread);
-        weekSpreadSum += weekSpread.size;
-        weekSpread.clear();
-        weekDate = timestamps[i];
-      }
-      if ((await this.addDays(sprintDate, 14)) <= new Date(timestamps[i])) {
-        dates.sprintSpread[sprintDate] = Array.from(sprintSpread);
-        sprintSpreadSum += sprintSpread.size;
-        sprintSpread.clear();
-        sprintDate = timestamps[i];
-      }
-      if ((await this.addDays(monthDate, 30)) <= new Date(timestamps[i])) {
-        dates.monthSpread[monthDate] = Array.from(monthSpread);
-        monthSpreadSum += monthSpread.size;
-        monthSpread.clear();
-        monthDate = timestamps[i];
-      }
-    }
-    // after all, add the computed numbers to the dates object
-    dates.daySpreadSum = daySpreadSum;
-    dates.weekSpreadSum = weekSpreadSum;
-    dates.sprintSpreadSum = sprintSpreadSum;
-    dates.monthSpreadSum = monthSpreadSum;
-    dates.days = Object.keys(dates.daySpread).length;
-    dates.weeks = Object.keys(dates.weekSpread).length;
-    dates.sprints = Object.keys(dates.sprintSpread).length;
-    dates.months = Object.keys(dates.monthSpread).length;
-    return dates;
   }
 }
